Updating Tree Node when new folder created
After a node has been expanded a second time the tree remembers children
of first time expansion and shows remebered list in tree.
New created folder can not be detected in this way.

Check if node has already expanded in past is removed. Every time tree
element is expanded children are listed new.

The change has been done by knowing that this slows down the system.
But the programmer decided that it is no use case to expan the same tree
element more than one time.
And in usual case the updating is still fast enough even if location is on
a server.